+++ routes/customers.router.js
@@ -3,7 +3,13 @@ import { Router } from 'express'
 
 // importo las funciones controladoras de la ruta
 
-import { pagar, getSorteoBuyers, getRifasBuyers, comprarSorteo,comprarRifa, updateCustomer, deleteCustomer, getCustomer, searchTicket, showFormBuy, getPriceDolar } from '../controllers/customers.controller.js'
+import {
+    getSorteoBuyers,
+    getRifasBuyers,
+    comprarSorteo,
+    comprarRifa,
+    searchTicket,
+ } from '../controllers/customers.controller.js'
 
 const router = Router()
 
@@ -14,21 +20,10 @@ router.get('/coindraw/getRifasBuyers', getRifasBuyers)
 
 router.get('/coindraw/searchTicket/:id_compra', searchTicket)
 
-router.get('/coindraw/getCustomer/:id', getCustomer)
 
 router.post('/coindraw/comprarSorteo', comprarSorteo)
 
 router.post('/coindraw/comprarRifa', comprarRifa)
-
-router.get('/coindraw/pagar', pagar)
-
-router.put('/coindraw/updateCustomer/:id', updateCustomer)
-
-router.delete('/coindraw/deleteCustomer/:id', deleteCustomer)
-
-router.get('/coindraw/comprar', showFormBuy)
-
-router.get('/coindraw/getDolar', getPriceDolar)
 
 
 // exporto las rutas para que se usen desde index.js
